refactor(session): split 733-line session.ts into modular architecture
Split monolithic session.ts into focused modules:
- validation.ts: Security validation (validateGitHubUrl, validateBranchName, shellEscape)
- types.ts: Shared TypeScript interfaces (ActiveSession)
- state.ts: Shared state management (activeSessions Map)
- constants.ts: Configuration constants (MODAL_APP_NAME, SNAPSHOT_INTERVAL_MS)
- snapshots.ts: Snapshot management (takeSnapshot, startSnapshotScheduler)
- sandbox.ts: Sandbox lifecycle (setupSandboxFromScratch, configureOpencode, startOpencodeServer, writeSecretsEnvFile)
- auth.ts: GitHub auth & git configuration (getGitHubAccount, getUserInfo, configureGitAuth)
- health.ts: Health monitoring (handleSessionDeath, startHealthMonitoring)
- index.ts: Core session functions and module re-exports

Benefits:
- Single responsibility per module
- Easier testing and maintenance
- Clear dependency structure
- Backward compatible via index.ts re-exports
- Health monitoring now explicitly started

Updated imports in:
- src/routes/api/projects..sessions.ts
- src/routes/api/projects..sessions..ts

All linting checks pass.
