+++ src/server/session/constants.ts
@@ -1,0 +1,5 @@
+export const MODAL_APP_NAME = process.env.MODAL_APP_NAME || "gondola";
+
+export const SNAPSHOT_INTERVAL_MS = Number(
+	process.env.SNAPSHOT_INTERVAL_MS || "300000",
+); // 5 min default
