+++ src/server/session/state.ts
@@ -1,0 +1,3 @@
+import type { ActiveSession } from "./types";
+
+export const activeSessions = new Map<string, ActiveSession>();
