+++ src/server/session/types.ts
@@ -1,0 +1,6 @@
+import type { Sandbox } from "modal";
+
+export interface ActiveSession {
+	sandbox: Sandbox;
+	snapshotInterval: ReturnType<typeof setInterval>;
+}
